requirements.txt
celery==5.3.1
django-cors-headers==4.2.0
django-filter==23.2
django-health-check==3.17.0
django==4.2.4
djangorestframework-api-key==2.3.0
djangorestframework-camel-case==1.4.2
djangorestframework==3.14.0
psycopg2==2.9.5
redis==4.6.0
uwsgi==2.0.22
black==23.7.0
coverage==7.2.7
django-extensions==3.2.3
flake8==6.1.0
freezegun==1.2.2
isort==5.12.0
mock==5.1.0
